feat!:  refactoring image generation with dev deps, tests and other stuff
