+++ pyproject.toml
@@ -10,8 +10,11 @@ repository = "https://github.com/ppp-one/cabaret"
 readme = "README.md"
 license = { text = "MIT License" }
 requires-python = ">=3.9"
-dependencies = ["numpy", "astropy"]
+dependencies = ["numpy", "astropy", "astroquery"]
 
 [build-system]
 requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
+
+[project.optional-dependencies]
+dev = ["black", "pytest"]
