+++ src/cabaret/__init__.py
@@ -1,0 +1,2 @@
+from cabaret.image import generate_image
+from cabaret.camera import Camera
