+++ src/cabaret/camera.py
@@ -6,9 +6,7 @@ class Camera:
     device_number: int = 0
     width: int = 1000
     height: int = 1000
-    numx: int = 1000
-    numy: int = 1000
-    maxadu: int = 2**16
+    max_adu: int = 2**16
     bin_x: int = 1
     bin_y: int = 1
     gain: float = 1.0
@@ -16,7 +14,6 @@ class Camera:
     plate_scale: float = 0.35
     well_depth: int = 2**16
     bias: int = 200
-    max_adu: int = 2**16
     read_noise: float = 4.0
     dark_current: float = 10.0
     sky_background: float = 10.0
